model_definitions/22c8a3b4-4909-4ed6-aee7-55978c1b2138/model_modules/requirements.txt
lightgbm==2.3.0
pypmml==0.9.7
sklearn-pandas==2.0.0
sklearn-pmml-model==0.0.15
sklearn2pmml==0.61.0